Build Gram via CSC x CSR sparse accumulation
Replace the pair HashMap + parallel reduce with a per-item dense
accumulator over a CSC view of X, emitted in CSR layout (u32 indices,
separate diagonal). Each row is summed in a fixed order, so P is
bit-identical across runs and thread counts; duplicates are summed.

+++ src/lib.rs
@@ -32,7 +32,7 @@ fn solve_slim<'py>(
     n_threads: Option<usize>,
 ) -> PyResult<CscArrays<'py>> {
     let data: Vec<f64> = data.as_slice()?.to_vec();
-    let indices: Vec<usize> = indices.as_slice()?.iter().map(|&x| x as usize).collect();
+    let indices: Vec<u32> = indices.as_slice()?.iter().map(|&x| x as u32).collect();
     let indptr: Vec<usize> = indptr.as_slice()?.iter().map(|&x| x as usize).collect();
     let params = SlimParams {
         lambd,
